requirements.txt
dj-database-url==0.5.0
django-debug-toolbar==3.1.1
  - Django [required: >=2.2, installed: 3.1.3]
    - asgiref [required: >=3.2.10,<4, installed: 3.3.0]      
    - pytz [required: Any, installed: 2020.4]
    - sqlparse [required: >=0.2.2, installed: 0.4.1]
  - sqlparse [required: >=0.2.0, installed: 0.4.1]
django-widget-tweaks==1.4.8
Markdown==3.3.3
python-decouple==3.3
whitenoise==5.2.0